React Routes and split into components
