+++ frontend/src/components/Secret.js
@@ -1,0 +1,4 @@
+import React, { useState } from 'react'
+import '../index.css'
+
+export const Secret = () => {}
